clochain-server/requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.32.1
python-dotenv==1.0.1
python-jose==3.3.0
pydantic==2.9.2
SQLAlchemy==2.0.36
qrcode==7.4.2
Pillow==11.0.0
eth-account==0.13.4
